+++ src/main.d.ts
@@ -1,58 +1,118 @@
 import { CustomError, ErrorName } from 'error-custom-class'
 
-type InvalidProps = 'name' | 'message' | 'stack' | 'errors' | 'cause'
-
 /**
- * Options passed to error constructors' second argument.
- */
-export interface Options {
-  /**
-   * Set as error properties.
-   *
-   * @example
-   * ```js
-   * const error = new InputError('Could not read the file.', {
-   *   props: { filePath: '/path' },
-   * })
-   * console.log(error.filePath) // '/path'
-   * ```
-   */
-  readonly props?: { [propName in InvalidProps]?: never } & {
-    [propName: string | symbol]: unknown
-  }
-
-  /**
-   * URL where users should report errors.
-   *
-   * @example
-   * ```js
-   * throw new InputError('Could not read the file.', {
-   *   bugs: 'https://github.com/my-name/my-project/issues',
-   * })
-   * // InputError: Could not read the file.
-   * // Please report this bug at: https://github.com/my-name/my-project/issues
-   * ```
-   */
-  readonly bugs?: string | URL
-}
-
-/**
- * Each known error class's `Error` ancestor class must be typed as `BaseError`.
+ * Each `custom` class's parent must be typed as `BaseError`.
  * The type parameter must match the class's name.
  *
  * @example
  * ```js
- * export class InputError extends (Error as BaseError<"InputError">) {}
+ * modernErrors({
+ *   TestError: { custom: class extends (Error as BaseError<'TestError'>) {} },
+ *   UnknownError: {},
+ * })
  * ```
  */
-export type BaseError<BaseErrorName extends ErrorName> = typeof CustomError<
-  BaseErrorName,
-  Options
+export type BaseError<BaseErrorName extends ErrorName> =
+  typeof CustomError<BaseErrorName>
+
+/**
+ * Class-specific options
+ */
+type ClassOptions = {
+  /**
+   * [Custom class](#custom-logic) to add any methods, `constructor` or
+   * properties.
+   * It must `extend` from
+   * [`Error`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error).
+   *
+   * @example
+   * ```js
+   * const { InputError, UnknownError, AnyError } = modernErrors({
+   *   // `*.custom` applies to a single class
+   *   InputError: {
+   *     custom: class extends Error {
+   *       constructor(message, options) {
+   *         // Modifying `message` or `options` should be done before `super()`
+   *         message += message.endsWith('.') ? '' : '.'
+   *
+   *         // `super()` should be called with both arguments
+   *         super(message, options)
+   *
+   *         // `name` is automatically added, so this is not necessary
+   *         // this.name = 'InputError'
+   *       }
+   *     },
+   *   },
+   *
+   *   // `AnyError.custom` applies to all classes
+   *   AnyError: {
+   *     custom: class extends Error {
+   *       isUserInput() {
+   *         return this.message.includes('user')
+   *       }
+   *     },
+   *   },
+   *
+   *   UnknownError: {},
+   * })
+   *
+   * const error = new InputError('Wrong user name')
+   * console.log(error.message) // 'Wrong user name.'
+   * console.log(error.isUserInput()) // true
+   * ```
+   */
+  readonly custom?: BaseError<ErrorName>
+}
+
+/**
+ * Error classes definitions. Object where:
+ *   - Each key is the class name, e.g. `InputError`.
+ *     One of the classes must be named `UnknownError`.
+ *   - Each value is an object with the class options.
+ */
+type Definitions = {
+  readonly [ErrorNameArg in ErrorName]: ClassOptions
+} & {
+  readonly UnknownError: ClassOptions
+}
+
+type ClassNames<DefinitionsArg extends Definitions> = Exclude<
+  keyof DefinitionsArg & ErrorName,
+  'AnyError'
 >
 
 /**
- * Base class of all `ErrorClasses` passed to `modernErrors()`.
- * Contains all the features.
+ * Error class returned by `modernErrors()`
+ */
+type ReturnErrorClass<
+  DefinitionsArg extends Definitions,
+  ErrorNameArg extends ErrorName,
+  CustomOption = DefinitionsArg[ErrorNameArg]['custom'],
+> = CustomOption extends BaseError<ErrorName>
+  ? CustomOption extends BaseError<ErrorNameArg>
+    ? CustomOption
+    : never
+  : BaseError<ErrorNameArg>
+
+/**
+ * All error classes returned by `modernErrors()`
+ */
+type ReturnErrorClasses<DefinitionsArg extends Definitions> = {
+  [ErrorNameArg in ClassNames<DefinitionsArg>]: ReturnErrorClass<
+    DefinitionsArg,
+    ErrorNameArg
+  >
+}
+
+type AnyKnownClass<
+  DefinitionsArg extends Definitions,
+  ErrorNameArg extends ErrorName = ClassNames<DefinitionsArg>,
+> = ErrorNameArg extends any
+  ? ReturnErrorClass<DefinitionsArg, ErrorNameArg>
+  : never
+
+/**
+ * Base class of the `ErrorClasses` passed to `modernErrors()`.
  *
  * @example
  * ```js
@@ -64,50 +124,49 @@ export type BaseError<BaseErrorName extends ErrorName> = typeof CustomError<
  * }
  * ```
  */
-type AnyError<ErrorClasses extends BaseError<ErrorName>> = ErrorClasses & {
-  /**
-   * Normalizes invalid errors and assigns the `UnknownError` class to
-   * _unknown_ errors. This should wrap each main function.
-   *
-   * @example
-   * ```js
-   * export const main = async function (filePath) {
-   *   try {
-   *     return await readContents(filePath)
-   *   } catch (error) {
-   *     throw AnyError.normalize(error)
-   *   }
-   * }
-   * ```
-   */
-  normalize(error: unknown): InstanceType<ErrorClasses>
-}
+type AnyError<DefinitionsArg extends Definitions> =
+  AnyKnownClass<DefinitionsArg> & {
+    /**
+     * Normalizes invalid errors and assigns the `UnknownError` class to
+     * _unknown_ errors. This should wrap each main function.
+     *
+     * @example
+     * ```js
+     * export const main = async function (filePath) {
+     *   try {
+     *     return await readContents(filePath)
+     *   } catch (error) {
+     *     throw AnyError.normalize(error)
+     *   }
+     * }
+     * ```
+     */
+    normalize(error: unknown): InstanceType<AnyKnownClass<DefinitionsArg>>
+  }
 
 /**
- * Changes the error classes' base class from `Error` to `AnyError`.
- * One of the error classes must be named `UnknownError`.
+ * Creates and returns error classes.
+ * Also returns their base class `AnyError`.
  *
  * @example
  * ```js
- * // Custom error classes
- * export class InputError extends Error {}
- * export class AuthError extends Error {}
- * export class DatabaseError extends Error {}
- * export class UnknownError extends Error {}
- *
- * // Base error class
- * export const AnyError = modernErrors([
+ * export const {
+ *   // Custom error classes
  *   InputError,
  *   AuthError,
  *   DatabaseError,
  *   UnknownError,
- * ])
+ *   // Base error class
+ *   AnyError,
+ * } = modernErrors({
+ *   // Custom error classes definitions
+ *   InputError: {},
+ *   AuthError: {},
+ *   DatabaseError: {},
+ *   UnknownError: {},
+ * })
  * ```
  */
-export default function modernErrors<
-  ErrorClasses extends BaseError<ErrorName>[],
->(
-  ErrorClasses: ErrorClasses,
-): BaseError<'UnknownError'> extends ErrorClasses[number]
-  ? AnyError<ErrorClasses[number]>
-  : never
+export default function modernErrors<DefinitionsArg extends Definitions>(
+  definitions: DefinitionsArg,
+): ReturnErrorClasses<DefinitionsArg> & { AnyError: AnyError<DefinitionsArg> }