+++ src/any/args.js
@@ -14,7 +14,9 @@ import isPlainObj from 'is-plain-obj'
 //     - Safer against injections
 export const normalizeConstructorArgs = function ({
   opts = {},
-  ErrorClasses,
+  ErrorClasses: {
+    UnknownError: { ErrorClass: UnknownError },
+  },
   AnyError,
   isAnyError,
 }) {
@@ -30,18 +32,7 @@ export const normalizeConstructorArgs = function ({
     )
   }
 
-  const UnknownError = getUnknownError(ErrorClasses)
   return normalizeCause({ opts, UnknownError, AnyError, isAnyError })
-}
-
-const getUnknownError = function ({ UnknownError }) {
-  if (UnknownError === undefined) {
-    throw new Error(
-      '"AnyError.class()" must be called before "new AnyError()".',
-    )
-  }
-
-  return UnknownError.ErrorClass
 }
 
 // `new AnyError()` does not make sense without a `cause`, so we validate it