+++ src/plugins/class_opts.js
@@ -1,7 +1,5 @@
 import isPlainObj from 'is-plain-obj'
 
-import { validateCustomUnknown } from '../subclass/unknown.js'
-
 import { mergePluginsOpts } from './merge.js'
 import { normalizePluginOpts } from './normalize.js'
 
@@ -63,4 +61,22 @@ export const getClassOpts = function ({
     normalizePluginOpts(classOptsB, plugin)
   })
   return { custom, classOpts: classOptsB }
+}
+
+// UnknownError cannot have a `custom` class:
+//  - This ensures its constructor does not throw
+//  - This discourages instantiating UnknownError directly, encouraging creating
+//    a separate error class for known system errors instead
+// We discourage extending or instantiating UnknownError but do not forbid it
+// since:
+//  - Users might not know instantiating would throw until runtime, which is
+//    problematic in error handling logic
+//  - There could be some potential use cases, e.g. if a branch is never meant
+//    to happen unless some unknown bug happened
+const validateCustomUnknown = function (custom, className) {
+  if (custom !== undefined && className === 'UnknownError') {
+    throw new TypeError(
+      'Error option "custom" is not available with "UnknownError".',
+    )
+  }
 }