+++ src/plugins/static.js
@@ -1,4 +1,4 @@
-import { requireUnknownError } from '../subclass/unknown.js'
+import { validateNonEmpty } from '../any/subclass.js'
 
 import { getErrorClasses } from './error_classes.js'
 import { normalizePluginOpts } from './normalize.js'
@@ -60,7 +60,7 @@ const callStaticMethods = function (
   { methodFunc, plugin, globalOpts, ErrorClasses, AnyError },
   ...args
 ) {
-  requireUnknownError(ErrorClasses)
+  validateNonEmpty(ErrorClasses)
   const options = normalizePluginOpts(globalOpts, plugin)
   return methodFunc(
     { options, AnyError, ErrorClasses: getErrorClasses(ErrorClasses) },