+++ test/subclass/unknown.js
@@ -3,10 +3,9 @@ import test from 'ava'
 import { createAnyError } from '../helpers/main.js'
 
 const AnyError = createAnyError()
-const InputError = AnyError.subclass('InputError')
 
 test('Require defining UnknownError before creating errors', (t) => {
-  t.throws(() => new InputError('test'))
+  t.throws(AnyError.subclass.bind(undefined, 'InputError'))
 })
 
 test('Require defining UnknownError before AnyError.normalize()', (t) => {
@@ -15,13 +14,6 @@ test('Require defining UnknownError before AnyError.normalize()', (t) => {
 
 test('Require defining UnknownError before plugin static methods', (t) => {
   t.throws(AnyError.getProp)
-})
-
-test('Allow defining UnknownError at the end', (t) => {
-  const TestAnyError = createAnyError()
-  const OtherInputError = TestAnyError.subclass('OtherInputError')
-  TestAnyError.subclass('UnknownError')
-  t.notThrows(() => new OtherInputError('test'))
 })
 
 test('Cannot use "custom" with UnknownError', (t) => {