+++ CHANGELOG.md
@@ -85,7 +85,7 @@ The [main function](README.md#modernerrorsplugins-options) now returns the base
 error class [`AnyError`](README.md#anyerror).
 
 [`AnyError.class(name)`](README.md#anyerrorclassname-options) must be used to
-create each error class. The first error class must now be named
+create each error class. One of them must now be named
 [`UnknownError`](README.md#unknown-errors).
 
 Before: