+++ CHANGELOG.md
@@ -133,6 +133,11 @@ error[methodName](...args, options[pluginName])
 The [`errors` option](README.md#aggregate-errors) can now be used to aggregate
 multiple errors into one, similarly to `new AggregateError(errors)`.
 
+### CLI plugin
+
+The [`modern-errors-cli` plugin](README.md#cli-errors) can now be used to handle
+CLI errors.
+
 ## Breaking changes
 
 ### Creating error classes